PlayerController, WorkersPatch Adjustments, Custom UI.Toggle
Added a grabber for our player object.

Adjusted worker stats to move faster.

creates a custom toggle that you pass an object through then it searches for all components and trys and finds the field/property you pass through it in any of them then sets it to true or false.

Suggestion, add a button that does the same? Maybe account for floats/ints as well.